Moved individual filename stuff to its own file
