+++ frontend/tests/unit/save-formats/N64/DexDrive.spec.js
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import seedrandom from 'seedrandom';
 import N64DexDriveSaveData from '@/save-formats/N64/DexDrive';
-import N64MempackSaveData from '@/save-formats/N64/Mempack';
+import N64GameSerialCodeUtil from '@/save-formats/N64/Components/GameSerialCodeUtil';
 import ArrayBufferUtil from '#/util/ArrayBuffer';
 
 const DIR = './tests/data/save-formats/n64/dexdrive';
@@ -227,11 +227,11 @@ describe('N64 - DexDrive save format', () => {
     expect(dexDriveSaveData.getSaveFiles()[0].noteName).to.equal('SMSM');
     expect(dexDriveSaveData.getSaveFiles()[0].noteNameExtension).to.equal('');
     expect(dexDriveSaveData.getSaveFiles()[0].comment).to.equal('100% Completed');
-    expect(dexDriveSaveData.getSaveFiles()[0].gameSerialCode).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_GAME_SERIAL_CODE);
-    expect(dexDriveSaveData.getSaveFiles()[0].publisherCode).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_PUBLISHER_CODE);
-    expect(dexDriveSaveData.getSaveFiles()[0].region).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_REGION_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[0].gameSerialCode).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_GAME_SERIAL_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[0].publisherCode).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_PUBLISHER_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[0].region).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_REGION_CODE);
     expect(dexDriveSaveData.getSaveFiles()[0].regionName).to.equal('Unknown region');
-    expect(dexDriveSaveData.getSaveFiles()[0].media).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_MEDIA_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[0].media).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_MEDIA_CODE);
     expect(ArrayBufferUtil.arrayBuffersEqual(dexDriveSaveData.getSaveFiles()[0].rawData, rawNoteArrayBuffer)).to.equal(true);
   });
 
@@ -252,11 +252,11 @@ describe('N64 - DexDrive save format', () => {
     expect(dexDriveSaveData.getSaveFiles()[0].noteName).to.equal('BK6');
     expect(dexDriveSaveData.getSaveFiles()[0].noteNameExtension).to.equal('SRAM');
     expect(dexDriveSaveData.getSaveFiles()[0].comment).to.equal('Save spot 3 contains a save with 63 jigsaws and 563 notes (time: 13:43:23)');
-    expect(dexDriveSaveData.getSaveFiles()[0].gameSerialCode).to.equal(N64MempackSaveData.BLACKBAG_CART_SAVE_GAME_SERIAL_CODE);
-    expect(dexDriveSaveData.getSaveFiles()[0].publisherCode).to.equal(N64MempackSaveData.BLACKBAG_CART_SAVE_PUBLISHER_CODE);
-    expect(dexDriveSaveData.getSaveFiles()[0].region).to.equal(N64MempackSaveData.BLACKBAG_CART_SAVE_REGION_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[0].gameSerialCode).to.equal(N64GameSerialCodeUtil.BLACKBAG_CART_SAVE_GAME_SERIAL_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[0].publisherCode).to.equal(N64GameSerialCodeUtil.BLACKBAG_CART_SAVE_PUBLISHER_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[0].region).to.equal(N64GameSerialCodeUtil.BLACKBAG_CART_SAVE_REGION_CODE);
     expect(dexDriveSaveData.getSaveFiles()[0].regionName).to.equal('Unknown region');
-    expect(dexDriveSaveData.getSaveFiles()[0].media).to.equal(N64MempackSaveData.BLACKBAG_CART_SAVE_MEDIA_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[0].media).to.equal(N64GameSerialCodeUtil.BLACKBAG_CART_SAVE_MEDIA_CODE);
     expect(ArrayBufferUtil.arrayBuffersEqual(dexDriveSaveData.getSaveFiles()[0].rawData, rawNote1ArrayBuffer)).to.equal(true);
 
     expect(dexDriveSaveData.getSaveFiles()[1].startingPage).to.equal(7);
@@ -264,11 +264,11 @@ describe('N64 - DexDrive save format', () => {
     expect(dexDriveSaveData.getSaveFiles()[1].noteName).to.equal('BK7');
     expect(dexDriveSaveData.getSaveFiles()[1].noteNameExtension).to.equal('SRAM');
     expect(dexDriveSaveData.getSaveFiles()[1].comment).to.equal('not working?');
-    expect(dexDriveSaveData.getSaveFiles()[1].gameSerialCode).to.equal(N64MempackSaveData.BLACKBAG_CART_SAVE_GAME_SERIAL_CODE);
-    expect(dexDriveSaveData.getSaveFiles()[1].publisherCode).to.equal(N64MempackSaveData.BLACKBAG_CART_SAVE_PUBLISHER_CODE);
-    expect(dexDriveSaveData.getSaveFiles()[1].region).to.equal(N64MempackSaveData.BLACKBAG_CART_SAVE_REGION_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[1].gameSerialCode).to.equal(N64GameSerialCodeUtil.BLACKBAG_CART_SAVE_GAME_SERIAL_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[1].publisherCode).to.equal(N64GameSerialCodeUtil.BLACKBAG_CART_SAVE_PUBLISHER_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[1].region).to.equal(N64GameSerialCodeUtil.BLACKBAG_CART_SAVE_REGION_CODE);
     expect(dexDriveSaveData.getSaveFiles()[1].regionName).to.equal('Unknown region');
-    expect(dexDriveSaveData.getSaveFiles()[1].media).to.equal(N64MempackSaveData.BLACKBAG_CART_SAVE_MEDIA_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[1].media).to.equal(N64GameSerialCodeUtil.BLACKBAG_CART_SAVE_MEDIA_CODE);
     expect(ArrayBufferUtil.arrayBuffersEqual(dexDriveSaveData.getSaveFiles()[1].rawData, rawNote2ArrayBuffer)).to.equal(true);
   });
 
@@ -290,11 +290,11 @@ describe('N64 - DexDrive save format', () => {
     expect(dexDriveSaveData.getSaveFiles()[0].noteNameExtension).to.equal('');
     expect(dexDriveSaveData.getSaveFiles()[0].comment).to.equal('101% Compleat, but I did not beat the game yet on this file.  '
       + 'Some stages have all 500 bananas collected, but others are missing some.  Try to find those if you want, but I\'ll try to update this file once I get those');
-    expect(dexDriveSaveData.getSaveFiles()[0].gameSerialCode).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_GAME_SERIAL_CODE);
-    expect(dexDriveSaveData.getSaveFiles()[0].publisherCode).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_PUBLISHER_CODE);
-    expect(dexDriveSaveData.getSaveFiles()[0].region).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_REGION_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[0].gameSerialCode).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_GAME_SERIAL_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[0].publisherCode).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_PUBLISHER_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[0].region).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_REGION_CODE);
     expect(dexDriveSaveData.getSaveFiles()[0].regionName).to.equal('Unknown region');
-    expect(dexDriveSaveData.getSaveFiles()[0].media).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_MEDIA_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[0].media).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_MEDIA_CODE);
     expect(ArrayBufferUtil.arrayBuffersEqual(dexDriveSaveData.getSaveFiles()[0].rawData, rawNoteArrayBuffer)).to.equal(true);
   });
 
@@ -343,11 +343,11 @@ describe('N64 - DexDrive save format', () => {
     expect(dexDriveSaveData.getSaveFiles()[2].noteName).to.equal('SMSM');
     expect(dexDriveSaveData.getSaveFiles()[2].noteNameExtension).to.equal('1');
     expect(dexDriveSaveData.getSaveFiles()[2].comment).to.equal('');
-    expect(dexDriveSaveData.getSaveFiles()[2].gameSerialCode).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_GAME_SERIAL_CODE);
-    expect(dexDriveSaveData.getSaveFiles()[2].publisherCode).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_PUBLISHER_CODE);
-    expect(dexDriveSaveData.getSaveFiles()[2].region).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_REGION_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[2].gameSerialCode).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_GAME_SERIAL_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[2].publisherCode).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_PUBLISHER_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[2].region).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_REGION_CODE);
     expect(dexDriveSaveData.getSaveFiles()[2].regionName).to.equal('Unknown region');
-    expect(dexDriveSaveData.getSaveFiles()[2].media).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_MEDIA_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[2].media).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_MEDIA_CODE);
     expect(ArrayBufferUtil.arrayBuffersEqual(dexDriveSaveData.getSaveFiles()[2].rawData, rawNote3ArrayBuffer)).to.equal(true);
 
     expect(dexDriveSaveData.getSaveFiles()[3].startingPage).to.equal(25); // There are 4 unused pages after the previous save. I assume something was deleted
@@ -355,11 +355,11 @@ describe('N64 - DexDrive save format', () => {
     expect(dexDriveSaveData.getSaveFiles()[3].noteName).to.equal('BKBK');
     expect(dexDriveSaveData.getSaveFiles()[3].noteNameExtension).to.equal('1');
     expect(dexDriveSaveData.getSaveFiles()[3].comment).to.equal('');
-    expect(dexDriveSaveData.getSaveFiles()[3].gameSerialCode).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_GAME_SERIAL_CODE);
-    expect(dexDriveSaveData.getSaveFiles()[3].publisherCode).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_PUBLISHER_CODE);
-    expect(dexDriveSaveData.getSaveFiles()[3].region).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_REGION_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[3].gameSerialCode).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_GAME_SERIAL_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[3].publisherCode).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_PUBLISHER_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[3].region).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_REGION_CODE);
     expect(dexDriveSaveData.getSaveFiles()[3].regionName).to.equal('Unknown region');
-    expect(dexDriveSaveData.getSaveFiles()[3].media).to.equal(N64MempackSaveData.GAMESHARK_ACTIONREPLAY_CART_SAVE_MEDIA_CODE);
+    expect(dexDriveSaveData.getSaveFiles()[3].media).to.equal(N64GameSerialCodeUtil.GAMESHARK_ACTIONREPLAY_CART_SAVE_MEDIA_CODE);
     expect(ArrayBufferUtil.arrayBuffersEqual(dexDriveSaveData.getSaveFiles()[3].rawData, rawNote4ArrayBuffer)).to.equal(true);
   });
